requirements.txt
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
plyer==2.1.0
pushbullet.py==0.12.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dotenv==1.0.1
python-magic==0.4.27
redis==5.2.1
requests==2.31.0
soupsieve==2.5
spotipy==2.25.0
urllib3==2.0.7
websocket-client==1.6.4
Werkzeug==3.1.3
